@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "npx next build",
+    "build": "next build",
     "start": "next start -p 8080",
     "lint": "next lint",
     "test": "jest",
@@ -23,13 +23,14 @@
     "apexcharts": "^4.3.0",
     "autoprefixer": "^10.4.20",
     "flatpickr": "^4.6.13",
-    "next": "14.2.35",
+    "next": "^14.2.35",
     "react": "^18.2.0",
     "react-apexcharts": "^1.7.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.3.5",
+    "react-intersection-observer": "^10.0.0",
     "swiper": "^11.2.0",
     "tailwind-merge": "^2.6.0"
   },
@@ -37,16 +38,16 @@
     "@eslint/eslintrc": "^3",
     "@svgr/webpack": "^8.1.0",
     "@swc/helpers": "^0.5.11",
-    "@testing-library/jest-dom": "^6.9.1",
+    "@tailwindcss/postcss": "^4.1.18",
     "@testing-library/react": "^16.3.1",
-    "@types/node": "24.8.1",
-    "@types/react": "18",
-    "@types/react-dom": "^19.2.3",
-    "eslint": "^8.57.0",
-    "jest": "^30.2.0",
-    "jest-environment-jsdom": "^30.2.0",
-    "eslint-config-next": "14.2.35",
-    "tailwindcss": "^3.4.1",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "@types/react-transition-group": "^4.4.12",
+    "eslint": "^9.0.0",
+    "eslint-config-next": "16.0.7",
+    "postcss": "^8",
+    "tailwindcss": "^4.0.0",
     "typescript": "^5"
   }
 }
